+++ pynicotine/shares.py
@@ -498,7 +498,6 @@ class Shares:
         self.share_dbs = {}
         self.public_rescanning = False
         self.buddy_rescanning = False
-        self.initiated_shares = False
 
         self.convert_shares()
         self.public_share_dbs = [
@@ -589,12 +588,6 @@ class Shares:
 
     def init_shares(self):
 
-        self.load_shares(self.share_dbs, self.public_share_dbs)
-        self.load_shares(self.share_dbs, self.buddy_share_dbs)
-
-        self.create_compressed_shares_message("normal")
-        self.create_compressed_shares_message("buddy")
-
         rescan_startup = self.config.sections["transfers"]["rescanonstartup"]
 
         # Rescan public shares if necessary
@@ -602,6 +595,8 @@ class Shares:
             if rescan_startup and not self.public_rescanning:
                 self.rescan_public_shares()
             else:
+                self.load_shares(self.share_dbs, self.public_share_dbs)
+                self.create_compressed_shares_message("normal")
                 self.compress_shares("normal")
 
         # Rescan buddy shares if necessary
@@ -609,9 +604,9 @@ class Shares:
             if rescan_startup and not self.buddy_rescanning:
                 self.rescan_buddy_shares()
             else:
+                self.load_shares(self.share_dbs, self.buddy_share_dbs)
+                self.create_compressed_shares_message("buddy")
                 self.compress_shares("buddy")
-
-        self.initiated_shares = True
 
     def convert_shares(self):
         """ Convert fs-based shared to virtual shared (pre 1.4.0) """
@@ -655,16 +650,51 @@ class Shares:
             if not reset_shares:
                 cls.load_shares(shares, dbs, reset_shares=True)
 
+    def file_is_shared(self, user, virtualfilename, realfilename):
+
+        log.add_transfer("Checking if file %(virtual_name)s with real path %(path)s is shared", {
+            "virtual_name": virtualfilename,
+            "path": realfilename
+        })
+
+        if not os.access(realfilename, os.R_OK):
+            log.add_transfer("Can't access file %(virtual_name)s with real path %(path)s, not sharing", {
+                "virtual_name": virtualfilename,
+                "path": realfilename
+            })
+            return False
+
+        folder, sep, file = virtualfilename.rpartition('\\')
+        shared = self.share_dbs.get("files")
+        bshared = self.share_dbs.get("buddyfiles")
+
+        if self.config.sections["transfers"]["enablebuddyshares"] and bshared is not None:
+            if user in (i[0] for i in self.config.sections["server"]["userlist"]):
+                for i in bshared.get(str(folder), ''):
+                    if file == i[0]:
+                        return True
+
+        if shared is not None:
+            for i in shared.get(str(folder), ''):
+                if file == i[0]:
+                    return True
+
+        log.add_transfer("Failed to share file %(virtual_name)s with real path %(path)s, since it wasn't found", {
+            "virtual_name": virtualfilename,
+            "path": realfilename
+        })
+        return False
+
     def add_file_to_shared(self, name):
         """ Add a file to the normal shares database """
 
         if not self.config.sections["transfers"]["sharedownloaddir"]:
             return
 
-        if not self.initiated_shares or self.public_rescanning:
-            return
+        shared = self.share_dbs.get("files")
 
-        shared = self.share_dbs["files"]
+        if shared is None:
+            return
 
         shareddirs = [path for _name, path in self.config.sections["transfers"]["shared"]]
         shareddirs.append(self.config.sections["transfers"]["downloaddir"])
@@ -695,19 +725,19 @@ class Shares:
             self.share_dbs["mtimes"][vdir] = os.path.getmtime(rdir)
             self.newnormalshares = True
 
-        if self.config.sections["transfers"]["enablebuddyshares"]:
-            self.add_file_to_buddy_shared(name)
-
     def add_file_to_buddy_shared(self, name):
         """ Add a file to the buddy shares database """
 
         if not self.config.sections["transfers"]["sharedownloaddir"]:
             return
 
-        if not self.initiated_shares or self.buddy_rescanning:
+        if not self.config.sections["transfers"]["enablebuddyshares"]:
             return
 
-        bshared = self.share_dbs["buddyfiles"]
+        bshared = self.share_dbs.get("buddyfiles")
+
+        if bshared is None:
+            return
 
         bshareddirs = [path for _name, path in self.config.sections["transfers"]["shared"]]
         bshareddirs += [path for _name, path in self.config.sections["transfers"]["buddyshared"]]
@@ -748,13 +778,13 @@ class Shares:
         if sharestype == "normal":
             self.compressed_shares_normal = slskmessages.SharedFileList(
                 None,
-                self.share_dbs["streams"]
+                self.share_dbs.get("streams")
             )
 
         elif sharestype == "buddy":
             self.compressed_shares_buddy = slskmessages.SharedFileList(
                 None,
-                self.share_dbs["buddystreams"]
+                self.share_dbs.get("buddystreams")
             )
 
     def get_compressed_shares_message(self, sharestype):
@@ -801,10 +831,11 @@ class Shares:
             ]
 
         for db in dbs:
-            try:
+            db_file = self.share_dbs.get(db)
+
+            if db_file is not None:
                 self.share_dbs[db].close()
-            except AttributeError:
-                continue
+                del self.share_dbs[db]
 
     def send_num_shared_folders_files(self):
         """ Send number publicly shared files to the server. """
@@ -815,17 +846,20 @@ class Shares:
             self.queue.append(slskmessages.SharedFoldersFiles(files, folders))
             return
 
-        shared_db = "files"
-        index_db = "fileindex"
-
         try:
+            shared = self.share_dbs.get("files")
+            index = self.share_dbs.get("fileindex")
+
+            if shared is None or index is None:
+                return
+
             try:
-                sharedfolders = len(self.share_dbs[shared_db])
-                sharedfiles = len(self.share_dbs[index_db])
+                sharedfolders = len(shared)
+                sharedfiles = len(index)
 
             except TypeError:
-                sharedfolders = len(list(self.share_dbs[shared_db]))
-                sharedfiles = len(list(self.share_dbs[index_db]))
+                sharedfolders = len(list(shared))
+                sharedfiles = len(list(index))
 
             self.queue.append(slskmessages.SharedFoldersFiles(sharedfolders, sharedfiles))