+++ pynicotine/transfers.py
@@ -37,7 +37,6 @@ import time
 
 from collections import defaultdict
 from collections import deque
-from hashlib import md5
 from time import sleep
 
 from pynicotine import slskmessages
@@ -53,7 +52,7 @@ from pynicotine.utils import write_file_and_backup
 class Transfer(object):
     """ This class holds information about a single transfer. """
 
-    __slots__ = "conn", "user", "realfilename", "filename", \
+    __slots__ = "conn", "user", "filename", \
                 "path", "req", "size", "file", "starttime", "lasttime", \
                 "offset", "currentbytes", "lastbytes", "speed", "timeelapsed", \
                 "timeleft", "timequeued", \
@@ -61,14 +60,13 @@ class Transfer(object):
                 "laststatuschange", "legacy_attempt"
 
     def __init__(
-        self, conn=None, user=None, realfilename=None, filename=None,
+        self, conn=None, user=None, filename=None,
         path=None, status=None, req=None, size=None, file=None, starttime=None,
         offset=None, currentbytes=None, speed=None, timeelapsed=None,
         timeleft=None, timequeued=None, requestconn=None,
         modifier=None, place=0, bitrate=None, length=None, iter=None, legacy_attempt=False
     ):
         self.user = user
-        self.realfilename = realfilename  # Sent as is to the user announcing what file we're sending
         self.filename = filename
         self.conn = conn
         self.path = path  # Used for ???
@@ -306,46 +304,6 @@ class Transfers:
         return False
 
     """ File Actions """
-
-    def file_is_shared(self, user, virtualfilename, realfilename):
-
-        log.add_transfer("Checking if file %(virtual_name)s with real path %(path)s is shared", {
-            "virtual_name": virtualfilename,
-            "path": realfilename
-        })
-
-        if not os.access(realfilename, os.R_OK):
-            log.add_transfer("Can't access file %(virtual_name)s with real path %(path)s, not sharing", {
-                "virtual_name": virtualfilename,
-                "path": realfilename
-            })
-            return False
-
-        folder, sep, file = virtualfilename.rpartition('\\')
-
-        if self.eventprocessor.shares.initiated_shares:
-
-            if self.config.sections["transfers"]["enablebuddyshares"] and \
-                    not self.eventprocessor.shares.buddy_rescanning:
-                if user in (i[0] for i in self.config.sections["server"]["userlist"]):
-                    bshared = self.eventprocessor.shares.share_dbs["buddyfiles"]
-
-                    for i in bshared.get(str(folder), ''):
-                        if file == i[0]:
-                            return True
-
-            if not self.eventprocessor.shares.public_rescanning:
-                shared = self.eventprocessor.shares.share_dbs["files"]
-
-                for i in shared.get(str(folder), ''):
-                    if file == i[0]:
-                        return True
-
-        log.add_transfer("Failed to share file %(virtual_name)s with real path %(path)s, since it wasn't found", {
-            "virtual_name": virtualfilename,
-            "path": realfilename
-        })
-        return False
 
     def get_file_size(self, filename):
 
@@ -644,9 +602,9 @@ class Transfers:
                     slskmessages.UploadDenied(conn=msg.conn.conn, file=msg.file, reason=limitmsg)
                 )
 
-            elif self.file_is_shared(user, filename_utf8, realpath):
+            elif self.eventprocessor.shares.file_is_shared(user, filename_utf8, realpath):
                 newupload = Transfer(
-                    user=user, filename=msg.file, realfilename=realpath,
+                    user=user, filename=msg.file,
                     path=os.path.dirname(realpath), status="Queued",
                     timequeued=time.time(), size=self.get_file_size(realpath), place=len(self.uploads)
                 )
@@ -795,7 +753,7 @@ class Transfers:
         # Do we actually share that file with the world?
         realpath = self.eventprocessor.shares.virtual2real(msg.file)
 
-        if not self.file_is_shared(user, msg.file, realpath):
+        if not self.eventprocessor.shares.file_is_shared(user, msg.file, realpath):
             return slskmessages.TransferResponse(None, 0, reason="File not shared", req=msg.req)
 
         # Is that file already in the queue?
@@ -836,7 +794,7 @@ class Transfers:
 
             response = slskmessages.TransferResponse(None, 0, reason="Queued", req=msg.req)
             newupload = Transfer(
-                user=user, filename=msg.file, realfilename=realpath,
+                user=user, filename=msg.file,
                 path=os.path.dirname(realpath), status="Queued",
                 timequeued=time.time(), size=self.get_file_size(realpath),
                 place=len(self.uploads)
@@ -850,7 +808,7 @@ class Transfers:
         response = slskmessages.TransferResponse(None, 1, req=msg.req, filesize=size)
 
         transferobj = Transfer(
-            user=user, realfilename=realpath, filename=msg.file,
+            user=user, filename=msg.file,
             path=os.path.dirname(realpath), status="Getting status",
             req=msg.req, size=size, place=len(self.uploads)
         )
@@ -1039,6 +997,7 @@ class Transfers:
             else:
                 f = None
                 try:
+                    from hashlib import md5
                     m = md5()
                     m.update((i.filename + i.user).encode('utf-8'))
 
@@ -1117,7 +1076,8 @@ class Transfers:
 
             try:
                 # Open File
-                f = open(i.realfilename, "rb")
+                realpath = self.eventprocessor.shares.virtual2real(i.filename)
+                f = open(realpath, "rb")
 
             except IOError as strerror:
                 log.add(_("Upload I/O error: %s"), strerror)
@@ -1566,20 +1526,17 @@ class Transfers:
 
         self.transfer_file(0, user, filename, path, transfer, size, bitrate, length)
 
-    def push_file(self, user, filename, realfilename, path="", transfer=None, size=None, bitrate=None, length=None, locally_queued=False):
-        if size is None:
-            size = self.get_file_size(realfilename)
+    def push_file(self, user, filename, path="", transfer=None, size=None, bitrate=None, length=None, locally_queued=False):
+        self.transfer_file(1, user, filename, path, transfer, size, bitrate, length, locally_queued)
 
-        self.transfer_file(1, user, filename, path, transfer, size, bitrate, length, realfilename, locally_queued)
-
-    def transfer_file(self, direction, user, filename, path="", transfer=None, size=None, bitrate=None, length=None, realfilename=None, locally_queued=False):
+    def transfer_file(self, direction, user, filename, path="", transfer=None, size=None, bitrate=None, length=None, locally_queued=False):
 
         """ Get a single file. path is a local path. if transfer object is
         not None, update it, otherwise create a new one."""
 
         if transfer is None:
             transfer = Transfer(
-                user=user, filename=filename, realfilename=realfilename, path=path,
+                user=user, filename=filename, path=path,
                 status="Queued", size=size, bitrate=bitrate,
                 length=length
             )
@@ -1896,6 +1853,7 @@ class Transfers:
         i.conn = None
 
         self.eventprocessor.shares.add_file_to_shared(newname)
+        self.eventprocessor.shares.add_file_to_buddy_shared(newname)
         self.eventprocessor.statistics.append_stat_value("completed_downloads", 1)
 
         # Attempt to show notification and execute commands
@@ -2129,8 +2087,7 @@ class Transfers:
                 continue
 
             self.push_file(
-                user=user, filename=upload_candidate.filename,
-                realfilename=upload_candidate.realfilename, transfer=upload_candidate
+                user=user, filename=upload_candidate.filename, transfer=upload_candidate
             )
             return